+++ packages/learn-card-helpers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @learncard/helpers
 
+## 1.2.4
+
+### Patch Changes
+
+-   Updated dependencies [[`016b7edc231273aab962b89b4351a3e229fca025`](https://github.com/learningeconomy/LearnCard/commit/016b7edc231273aab962b89b4351a3e229fca025)]:
+    -   @learncard/types@5.11.3
+
 ## 1.2.3
 
 ### Patch Changes