+++ packages/plugins/lca-api-plugin/CHANGELOG.md
@@ -1,5 +1,16 @@
 # lca-api-plugin
 
+## 1.1.6
+
+### Patch Changes
+
+-   Updated dependencies [[`016b7edc231273aab962b89b4351a3e229fca025`](https://github.com/learningeconomy/LearnCard/commit/016b7edc231273aab962b89b4351a3e229fca025)]:
+    -   @learncard/types@5.11.3
+    -   @learncard/lca-api-client@1.1.6
+    -   @learncard/core@9.4.4
+    -   @learncard/init@2.2.6
+    -   @learncard/didkit-plugin@1.6.4
+
 ## 1.1.5
 
 ### Patch Changes