+++ packages/plugins/learn-cloud/CHANGELOG.md
@@ -1,5 +1,15 @@
 # learn-card-core
 
+## 2.3.6
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.4.4
+    -   @learncard/helpers@1.2.4
+    -   @learncard/learn-cloud-client@1.6.6
+    -   @learncard/didkit-plugin@1.6.4
+
 ## 2.3.5
 
 ### Patch Changes